refactor: 책임 분리를 위한 서비스 계층 도입
